+++ src/Discord.Net.Rest/API/Common/ApplicationCommandOption.cs
@@ -30,10 +30,10 @@ namespace Discord.API
         public Optional<bool> Autocomplete { get; set; }
 
         [JsonProperty("min_value")]
-        public Optional<int> MinValue { get; set; }
+        public Optional<double> MinValue { get; set; }
 
         [JsonProperty("max_value")]
-        public Optional<int> MaxValue { get; set; }
+        public Optional<double> MaxValue { get; set; }
 
         [JsonProperty("channel_types")]
         public Optional<ChannelType[]> ChannelTypes { get; set; }
@@ -54,8 +54,8 @@ namespace Discord.API
 
             Required = cmd.IsRequired ?? Optional<bool>.Unspecified;
             Default = cmd.IsDefault ?? Optional<bool>.Unspecified;
-            MinValue = cmd.MinValue ?? Optional<int>.Unspecified;
-            MaxValue = cmd.MaxValue ?? Optional<int>.Unspecified;
+            MinValue = cmd.MinValue ?? Optional<double>.Unspecified;
+            MaxValue = cmd.MaxValue ?? Optional<double>.Unspecified;
 
             Name = cmd.Name;
             Type = cmd.Type;
@@ -74,8 +74,8 @@ namespace Discord.API
             Required = option.Required ?? Optional<bool>.Unspecified;
 
             Default = option.Default ?? Optional<bool>.Unspecified;
-            MinValue = option.MinValue ?? Optional<int>.Unspecified;
-            MaxValue = option.MaxValue ?? Optional<int>.Unspecified;
+            MinValue = option.MinValue ?? Optional<double>.Unspecified;
+            MaxValue = option.MaxValue ?? Optional<double>.Unspecified;
 
             ChannelTypes = option.ChannelTypes?.ToArray() ?? Optional<ChannelType[]>.Unspecified;
