+++ src/Discord.Net.WebSocket/Entities/Interaction/SocketBaseCommand/SocketApplicationCommandOption.cs
@@ -26,10 +26,10 @@ namespace Discord.WebSocket
         public bool? IsRequired { get; private set; }
 
         /// <inheritdoc/>
-        public int? MinValue { get; private set; }
+        public double? MinValue { get; private set; }
 
         /// <inheritdoc/>
-        public int? MaxValue { get; private set; }
+        public double? MaxValue { get; private set; }
 
         /// <summary>
         ///     Choices for string and int types for the user to pick from.
